v2
Added: if package file does not contain ‘->’ program will terminate.

Cleaned code, improved comments, prepared for submission to BT
