+++ src/packages.txt
@@ -4,3 +4,4 @@ textui -> runner framework
 awtui -> runner
 runner -> framework
 extensions -> framework
+hello
